django/requirements.txt
dj-database-url==0.3.0
Django==1.8.7
djangorestframework==3.3.1
psycopg2==2.6.1
python-decouple==3.0
python-social-auth==0.2.13
rest-social-auth==0.2.0
